refactor: extract creating article logic into controller
