+++ src/controller/__init__.py
@@ -1,0 +1,1 @@
+from .article import create_article_controller, BodyCreateArticle
